+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * `firewalld` permanent type fix
 * `pause` validation fix for `minutes` and `seconds`
 * `user` `groups` correction until Ansible 2.3
+* `gem` module now works and does not conflict with the Ruby keyword
 
 # 1.0.6 - 14 October 2016
 * Allow sharing variables between tasks (#29)